+++ Tweaks/AutoOpenRecipe.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using Dalamud.Game.ClientState.Conditions;
+using Dalamud.Utility;
 using Dalamud.Utility.Signatures;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using FFXIVClientStructs.FFXIV.Client.Game.Character;
@@ -15,8 +16,8 @@ namespace HaselTweaks.Tweaks;
 
 public unsafe partial class AutoOpenRecipe : Tweak
 {
-    public override string Name => "Auto-open Recipe (Experimental)";
-    public override string Description => "When a quest step requires an item which is craftable and you have materials in your inventory, it will automatically open the recipe.";
+    public override string Name => "Auto-open Recipe";
+    public override string Description => "When a daily quest objective requires an item which is craftable and you have materials in your inventory, it will automatically open the recipe.";
 
     private ExcelSheet<Quest> questSheet = null!;
     private ExcelSheet<Recipe> recipeSheet = null!;
@@ -28,6 +29,80 @@ public unsafe partial class AutoOpenRecipe : Tweak
     private QuestManager* questManager;
     private RecipeNote* recipeNote;
     private PlayerState* playerState;
+
+    // for older quests that don't return the item id in GetTodoArgs
+    private readonly record struct QuestTodo(ushort QuestId, byte TodoIndex, string ScriptArgName = "RITEM1");
+    private readonly QuestTodo[] QuestTodos = new QuestTodo[] {
+        // Ixal
+        new(1494, 3),
+        new(1495, 3),
+        new(1496, 2),
+        new(1497, 2),
+        new(1504, 2),
+        new(1505, 2),
+        new(1506, 2),
+        new(1507, 2),
+        new(1508, 2),
+        new(1514, 2),
+        new(1515, 3),
+        new(1516, 3),
+        new(1517, 2),
+        new(1518, 3),
+        new(1498, 4),
+        new(1499, 4),
+        new(1500, 2),
+        new(1501, 2),
+        new(1502, 3),
+        new(1503, 2),
+        new(1509, 3),
+        new(1510, 4),
+        new(1511, 4),
+        new(1512, 3),
+        new(1513, 3),
+        new(1519, 3),
+        new(1520, 3),
+        new(1521, 4),
+        new(1522, 4),
+        new(1523, 3),
+        new(1566, 2),
+        new(1567, 2),
+        new(1568, 2),
+        new(1487, 2, "RITEM2"),
+        new(1487, 3, "RITEM2"),
+        new(1488, 3),
+        new(1489, 4),
+        new(1491, 2),
+        new(1493, 1),
+        new(1493, 12, "RITEM3"),
+
+        // Moogle
+        new(2320, 3),
+        new(2322, 5),
+        new(2324, 4),
+        new(2325, 5),
+        new(2326, 7),
+        new(2290, 1),
+        new(2291, 1),
+        new(2292, 3),
+        new(2293, 3),
+        new(2294, 3),
+        new(2296, 1),
+        new(2298, 1),
+        new(2299, 1),
+        new(2300, 1),
+        new(2301, 1),
+        new(2303, 1),
+        new(2304, 2),
+        new(2305, 4),
+        new(2307, 1),
+        new(2310, 2),
+        new(2311, 3),
+        new(2313, 3),
+        new(2314, 2),
+        new(2316, 2),
+        new(2317, 1),
+        new(2318, 5),
+    };
 
     public override void Setup()
     {
@@ -49,7 +124,8 @@ public unsafe partial class AutoOpenRecipe : Tweak
         var questId = *(ushort*)questData;
         Debug($"UpdateQuestWork({index}, {questId} / {*(byte*)(questData + 2)}, {a3}, {a4}, {a5})");
 
-        if (questId > 0 /*&& questManager->GetDailyQuestById(questId) != null*/) // TODO: make this an option? maybe an exclude list?
+        // DailyQuestWork gets updated before QuestWork, so we can check here if it's a daily quest
+        if (questId > 0 && questManager->GetDailyQuestById(questId) != null)
         {
             var quest = questSheet.GetRow((uint)questId | 0x10000);
             if (quest?.RowId == 0)
@@ -59,26 +135,42 @@ public unsafe partial class AutoOpenRecipe : Tweak
             }
 
             var sequence = *(byte*)(questData + 2);
-            var todoIndex = quest!.ToDoCompleteSeq.IndexOf(sequence);
-            var todoCount = quest.ToDoQty[todoIndex];
+            var todoOffset = quest!.ToDoCompleteSeq.IndexOf(sequence);
+            if (todoOffset < 0 || todoOffset >= quest.ToDoQty.Length)
+                goto originalUpdateQuestWork;
 
             var questEventHandler = EventFramework.Instance()->GetEventHandlerById(questId);
             var localPlayer = Control.Instance()->LocalPlayer;
-            if (questEventHandler != null && localPlayer != null)
-            {
-                for (var i = todoIndex; i < todoIndex + todoCount; i++) // TODO: not sure if this is correct
-                {
-                    uint numHave;
-                    uint numNeeded;
-                    uint itemId;
-                    GetTodoArgs(questEventHandler, localPlayer, i, &numHave, &numNeeded, &itemId);
-                    Debug($"TodoArgs #{i}: {numHave}/{numNeeded} of {itemId}");
+            if (questEventHandler == null && localPlayer == null)
+                goto originalUpdateQuestWork;
 
-                    if (numHave < numNeeded && itemId != 0)
+            var todoCount = quest.ToDoQty[todoOffset];
+            for (var todoIndex = todoOffset; todoIndex < todoOffset + todoCount; todoIndex++)
+            {
+                uint numHave, numNeeded, itemId;
+                GetTodoArgs(questEventHandler, localPlayer, todoIndex, &numHave, &numNeeded, &itemId);
+                Debug($"TodoArgs #{todoIndex}: {numHave}/{numNeeded} of {itemId}");
+
+                if (itemId == 0)
+                {
+                    foreach (var q in QuestTodos)
                     {
-                        OpenRecipe(itemId % 1000000);
-                        break;
+                        if (q.QuestId == questId && q.TodoIndex == todoIndex)
+                        {
+                            var scriptArgIndex = quest.ScriptInstruction
+                                .IndexOf(entry => entry.RawString == q.ScriptArgName);
+
+                            itemId = quest.ScriptArg.ElementAtOrDefault(scriptArgIndex);
+                            Debug($"Using fallback script value {itemId}");
+                            break;
+                        }
                     }
+                }
+
+                if (numHave < numNeeded && itemId != 0)
+                {
+                    OpenRecipe(itemId % 1000000, numNeeded);
+                    break;
                 }
             }
         }
@@ -91,7 +183,7 @@ public unsafe partial class AutoOpenRecipe : Tweak
     private readonly GetTodoArgsDelegate GetTodoArgs = null!;
     private delegate void GetTodoArgsDelegate(EventHandler* questEventHandler, BattleChara* localPlayer, int i, uint* numHave, uint* numNeeded, uint* itemId);
 
-    private void OpenRecipe(uint resultItemId)
+    private void OpenRecipe(uint resultItemId, uint amount)
     {
         if (agentRecipeNote->ActiveCraftRecipeId != 0 || Service.Condition[ConditionFlag.Crafting] || Service.Condition[ConditionFlag.Crafting40])
         {
@@ -100,14 +192,15 @@ public unsafe partial class AutoOpenRecipe : Tweak
         }
 
         var craftType = GetCurrentCraftType();
-        var recipe = recipeSheet?.FirstOrDefault(row => row?.ItemResult.Row == resultItemId && (craftType == -1 || row.CraftType.Row == craftType), null);
+        var recipe = recipeSheet?.FirstOrDefault(row => row?.ItemResult.Row == resultItemId && row.CraftType.Row == craftType, null);
+        recipe ??= recipeSheet?.FirstOrDefault(row => row?.ItemResult.Row == resultItemId, null);
         if (recipe == null)
         {
             Warning($"Not opening Recipe for Item {resultItemId}: Recipe not found");
             return;
         }
 
-        if (!IngredientsAvailable(recipe))
+        if (!IngredientsAvailable(recipe, amount))
         {
             Warning($"Not opening Recipe for Item {resultItemId}: Required ingredients not available");
             return;
@@ -115,15 +208,15 @@ public unsafe partial class AutoOpenRecipe : Tweak
 
         Debug($"Requirements met, opening recipe {recipe.RowId}");
 
-        // TODO: check if recipe is already open?
         agentRecipeNote->AgentInterface.Hide();
 
-        // TODO: make config option which one to use?
-        // agentRecipeNote->OpenRecipeByItemId(resultItemId);
-        agentRecipeNote->OpenRecipeByRecipeIdInternal(recipe.RowId);
+        if (recipe.CraftType.Row == craftType)
+            agentRecipeNote->OpenRecipeByRecipeIdInternal(recipe.RowId);
+        else
+            agentRecipeNote->OpenRecipeByItemId(resultItemId);
     }
 
-    private bool IngredientsAvailable(Recipe recipe)
+    private bool IngredientsAvailable(Recipe recipe, uint amount)
     {
         foreach (var ingredient in recipe.UnkData5)
         {
@@ -136,9 +229,14 @@ public unsafe partial class AutoOpenRecipe : Tweak
             if (item == null || item.ItemUICategory.Row == 59) // ignore Crystals
                 continue;
 
-            var count = inventoryManager->GetInventoryItemCount(itemId, false, false, false);
-            Debug($"Checking Ingredient #{ingredient.ItemIngredient}: need {ingredient.AmountIngredient}, have {count}");
-            if (count < ingredient.AmountIngredient)
+            var numHave = inventoryManager->GetInventoryItemCount(itemId, false, false, false); // Normal
+            numHave += inventoryManager->GetInventoryItemCount(itemId, true, false, false); // HQ
+
+            var numNeeded = ingredient.AmountIngredient * amount;
+
+            Debug($"Checking Ingredient #{ingredient.ItemIngredient}: need {numNeeded}, have {numHave}");
+
+            if (numHave < numNeeded)
                 return false;
         }
 
@@ -147,12 +245,13 @@ public unsafe partial class AutoOpenRecipe : Tweak
 
     private int GetCurrentCraftType()
     {
-        var craftType = 0;
+        var craftType = -1;
 
-        for (; craftType < craftTypeSheet.RowCount; craftType++)
+        for (var i = 0; i < craftTypeSheet.RowCount; i++)
         {
-            if (recipeNote->Jobs[craftType] == playerState->CurrentClassJobId)
+            if (recipeNote->Jobs[i] == playerState->CurrentClassJobId)
             {
+                craftType = i;
                 break;
             }
         }